+++ src/oph_workflow_engine.c
@@ -1353,11 +1353,45 @@ int oph_workflow_parallel_fco(oph_workflow * wf, int nesting_level)
 			if (!replied_num)
 				continue;
 
+			var.caller = i;
+
+			// Add the number of branches to workflow environment
+			size_t name_size = 2 + strlen(name) + strlen(OPH_WORKFLOW_COUNTER_SIZE);
+			char number_of_loops[name_size];
+			snprintf(number_of_loops, var_size, "%s_" OPH_WORKFLOW_COUNTER_SIZE, name);
+			if (!hashtbl_get(wf->vars, number_of_loops)) {
+				var.ivalue = 1;	// Non used
+				var.svalue = (char *) calloc(OPH_WORKFLOW_MIN_STRING, sizeof(char));
+				if (var.svalue)
+					snprintf(var.svalue, OPH_WORKFLOW_MIN_STRING, "%d", svalues_num);
+				if (!var.svalue) {
+					pmesg_safe(&global_flag, LOG_ERROR, __FILE__, __LINE__, "Memory error\n");
+					break;
+				}
+				svalue_size = strlen(var.svalue) + 1;
+				var_buffer = malloc(var_size + svalue_size);
+				if (!var_buffer) {
+					pmesg_safe(&global_flag, LOG_ERROR, __FILE__, __LINE__, "Memory error\n");
+					free(var.svalue);
+					break;
+				}
+				memcpy(var_buffer, (void *) &var, var_size);
+				memcpy(var_buffer + var_size, var.svalue, svalue_size);
+				if (hashtbl_insert_with_size(wf->vars, number_of_loops, var_buffer, var_size + svalue_size)) {
+					pmesg_safe(&global_flag, LOG_DEBUG, __FILE__, __LINE__, "Unable to store variable '%s' in environment of task '%s'. Maybe it already exists.\n", number_of_loops, wf->tasks[j].name);
+					free(var.svalue);
+					free(var_buffer);
+					break;
+				}
+				pmesg_safe(&global_flag, LOG_DEBUG, __FILE__, __LINE__, "Added variable '%s=%s' in environment of task '%s'.\n", number_of_loops, var.svalue, wf->tasks[j].name);
+				free(var.svalue);
+				free(var_buffer);
+			}
+
 			// In case no additional branch has to be created, simply add for-variable as task local variable to marked tasks
 			if (!new_branch_num) {
 				for (j = 0; j < wf->tasks_num; ++j)
 					if (wf->tasks[j].is_marked) {
-						var.caller = i;
 						if (ivalues)
 							var.ivalue = ivalues[0];
 						else
@@ -1389,12 +1423,7 @@ int oph_workflow_parallel_fco(oph_workflow * wf, int nesting_level)
 							free(var_buffer);
 							break;
 						}
-						if (svalues)
-							pmesg_safe(&global_flag, LOG_DEBUG, __FILE__, __LINE__, "Added variable '%s=%s' in environment of task '%s'.\n", name, var.svalue,
-								   wf->tasks[j].name);
-						else
-							pmesg_safe(&global_flag, LOG_DEBUG, __FILE__, __LINE__, "Added variable '%s=%d' in environment of task '%s'.\n", name, var.ivalue,
-								   wf->tasks[j].name);
+						pmesg_safe(&global_flag, LOG_DEBUG, __FILE__, __LINE__, "Added variable '%s=%s' in environment of task '%s'.\n", name, var.svalue, wf->tasks[j].name);
 						free(var.svalue);
 						free(var_buffer);
 					}
@@ -1531,7 +1560,6 @@ int oph_workflow_parallel_fco(oph_workflow * wf, int nesting_level)
 						k = 0;
 					else
 						k = 1 + (j - old_tasks_num) / replied_num;
-					var.caller = i;
 					if (ivalues)
 						var.ivalue = ivalues[k];
 					else